requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
langchain
langchain-openai
langchain-mistralai
langchain-google-genai
langchain-core
langchain-ollama==0.3.7
google-generativeai
mistralai
openai
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
langsmith
requests==2.31.0
websockets==12.0 